Refacto and improvements
